// orbito-frontend/src/core/api/generated/models/updateSubscriptionPlanCommand.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Orbito API
 * OpenAPI spec version: v1
 */
import type { BillingPeriodType } from './billingPeriodType';

export interface UpdateSubscriptionPlanCommand {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  amount?: number;
  /** @nullable */
  currency?: string | null;
  billingPeriodType?: BillingPeriodType;
  trialDays?: number;
  trialPeriodDays?: number;
  /** @nullable */
  featuresJson?: string | null;
  /** @nullable */
  limitationsJson?: string | null;
  isActive?: boolean;
  isPublic?: boolean;
  sortOrder?: number;
}